requirements-health.txt
attrs==25.4.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
-e git+ssh://git@github.com/Aliyev085/ETL_PROJECT_1.git@3b722dbc213da7b2289a9cb23362318f9e91a133#egg=etlserver
exceptiongroup==1.3.0
Flask==3.1.2
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
outcome==1.3.0.post0
pika==1.3.2
psutil==7.1.3
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.11
PySocks==1.7.1
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
selenium==4.36.0
selenium-stealth==1.0.6
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.9.0
Werkzeug==3.1.3
wsproto==1.2.0
